breaking up the code into pieces
