--- random.js
@@ -1,11 +1,0 @@
-const randomStringGenerator = (length = 52) => {
-  let result = '';
-  const chars =
-    'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
-  for (let i = 0; i < length; i++) {
-    result += chars.charAt(Math.floor(Math.random() * chars.length));
-  }
-  return result;
-};
-
-module.exports = randomStringGenerator;
